Added Custom Serializer Options for pcs4 datetime Moved MassTransit code to separate class
