+++ src/Equinor.ProCoSys.Completion.WebApi/appsettings.json
@@ -54,7 +54,7 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "MassTransit": "Debug",
+      "MassTransit": "Information",
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     },
